fix(sdk/analyze): unify tool-result stringify on the TS-faithful version
hotspots and patterns each carried a copy of stringifyToolResult. The TS
originals (recovered from a573de5~1) are byte-identical, but patterns' Rust
port diverged at the array catch-all: it JSON-stringified bare scalar blocks
(number/bool/null) that TS — and the hotspots port — skip, because such a
block is neither `typeof === 'object'` nor `typeof === 'string'`.

Move the TS-faithful version (hotspots') into analyze::util::stringify_tool_result
and call it from both, deleting the two copies. This realigns patterns with
the TS source and removes the drift permanently.

No observable change on real input: Claude/Codex tool_result.content is a bare
string or an array of typed content-block objects, never bare scalars, so the
corrected arm is unreachable in practice and no test exercises it. 992/992
workspace tests pass.
